Added Colombia Holidays (#55)
* Adding Colombia Holidays

Added the Colombia's holidays. There's a situation with some holidays due to the "Emiliany Law" they are moved to the following monday.

* Adding Colombia Holidays

Added the Colombia's holidays. There's a situation with some holidays due to the "Emiliany Law" they are moved to the following monday.

* Update Colombia.php

Deleted the line as suggested by @Nielsvanpach

* Update Colombia.php

* Added changes to Colombia.php and to the test file

- Suggestions from @luisprmat were incorporated.
- All the work of emiliani holiday generation was included into the emilianiHoliday method to achieve code readability
- The order of the holidays was updated to ensure that the next emiliani holiday that appears can be included at the end of the list.
- The test was updated because it's "colombian holidays" not "colombia holidays"